Update percent calculation

+++ src/Traq/Models/Milestone.php
@@ -142,9 +142,9 @@ class Milestone extends Model
 
         if (count($results)) {
             $counts[$this->id] = $results;
-            $counts[$this->id]['open_percent'] = (int) get_percent(($counts[$this->id]['open'] ?? 0) - ($counts[$this->id]['started'] ?? 0), $counts[$this->id]['total']);
-            $counts[$this->id]['started_percent'] = (int) get_percent($counts[$this->id]['started'] ?? 0, $counts[$this->id]['total']);
-            $counts[$this->id]['closed_percent'] = (int) get_percent($counts[$this->id]['closed'] ?? 0, $counts[$this->id]['total']);
+            $counts[$this->id]['open_percent'] = get_percent(($counts[$this->id]['open'] ?? 0) - ($counts[$this->id]['started'] ?? 0), $counts[$this->id]['total']);
+            $counts[$this->id]['started_percent'] = get_percent($counts[$this->id]['started'] ?? 0, $counts[$this->id]['total']);
+            $counts[$this->id]['closed_percent'] = get_percent($counts[$this->id]['closed'] ?? 0, $counts[$this->id]['total']);
         }
 
         // Return the requested count index.
